Добавляет управляемый компонент формы отправки комментария
