+++ src/components/offer-card/offer-card.jsx
