--- src/components/room/room.jsx
@@ -1,224 +1,0 @@
-import React from "react";
-import {Link} from "react-router-dom";
-import PropTypes from "prop-types";
-import {offersPropTypes} from "../../offers-prop-types";
-import {reviewsPropTypes} from "../../reviews-prop-types";
-import {getRatingInPercent, capitalize, formatDate} from "../../util/offer";
-import {CardType} from "../../const";
-import {nanoid} from "nanoid";
-import Card from "../card/card";
-
-const Room = (props) => {
-  const {offerId, offers, reviews, favorites} = props;
-  const room = offers.find((offer) => offer.id === offerId);
-  const {title, price, type, description, bedrooms, guests, rating, photos, options, owner, isPremial} = room;
-  const isFavorite = favorites.includes(offerId);
-  const roomReviews = reviews.find((item) => item.propertyId === offerId).reviews;
-  const isAnyReviews = roomReviews.length !== 0 ? true : false;
-  let descriptionSentences = description.split(`. `);
-  descriptionSentences[descriptionSentences.length - 1] = descriptionSentences[descriptionSentences.length - 1].replace(`.`, ``);
-  const neighbourhoodOffers = offers.slice(0, 3); // временно
-
-  return (
-    <div className="page">
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <Link className="header__logo-link" to={`/`}>
-                <img className="header__logo" src="/img/logo.svg" alt="6 cities logo" width="81" height="41" />
-              </Link>
-            </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <a className="header__nav-link header__nav-link--profile" href="#">
-                    <div className="header__avatar-wrapper user__avatar-wrapper">
-                    </div>
-                    <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
-                  </a>
-                </li>
-              </ul>
-            </nav>
-          </div>
-        </div>
-      </header>
-
-      <main className="page__main page__main--property">
-        <section className="property">
-          <div className="property__gallery-container container">
-            <div className="property__gallery">
-              {photos.length === 0 ? `` : photos.map((photo) =>
-                <div key={nanoid()} className="property__image-wrapper">
-                  <img className="property__image" src={photo.src} alt={photo.description} />
-                </div>
-              )}
-            </div>
-          </div>
-          <div className="property__container container">
-            <div className="property__wrapper">
-              {isPremial ? <div className="property__mark">
-                <span>Premium</span>
-              </div> : ``}
-              <div className="property__name-wrapper">
-                <h1 className="property__name">{title}</h1>
-                <button className={`property__bookmark-button ${isFavorite ? `property__bookmark-button--active` : ``} button`} type="button">
-                  <svg className="property__bookmark-icon" width="31" height="33">
-                    <use xlinkHref="#icon-bookmark"></use>
-                  </svg>
-                  <span className="visually-hidden">{`${isFavorite ? `In` : `To`} bookmarks`}</span>
-                </button>
-              </div>
-              <div className="property__rating rating">
-                <div className="property__stars rating__stars">
-                  <span style={{width: `${getRatingInPercent(rating)}%`}}></span>
-                  <span className="visually-hidden">Rating</span>
-                </div>
-                <span className="property__rating-value rating__value">{rating}</span>
-              </div>
-              <ul className="property__features">
-                <li className="property__feature property__feature--entire">
-                  {capitalize(type)}
-                </li>
-                <li className="property__feature property__feature--bedrooms">
-                  {`${bedrooms} Bedrooms`}
-                </li>
-                <li className="property__feature property__feature--adults">
-                  {`Max ${guests} adults`}
-                </li>
-              </ul>
-              <div className="property__price">
-                <b className="property__price-value">&euro;{price}</b>
-                <span className="property__price-text">&nbsp;night</span>
-              </div>
-              <div className="property__inside">
-                <h2 className="property__inside-title">What&apos;s inside</h2>
-                <ul className="property__inside-list">
-                  {options.map((option) =>
-                    <li key={nanoid()} className="property__inside-item">{option}</li>
-                  )}
-                </ul>
-              </div>
-              <div className="property__host">
-                <h2 className="property__host-title">Meet the host</h2>
-                <div className="property__host-user user">
-                  <div className={`property__avatar-wrapper ${owner.isHighRated ? `property__avatar-wrapper--pro` : ``} user__avatar-wrapper`}>
-                    <img className="property__avatar user__avatar" src={owner.avatar} width="74" height="74" alt="Host avatar" />
-                  </div>
-                  <span className="property__user-name">
-                    {owner.name}
-                  </span>
-                </div>
-                <div className="property__description">
-                  {descriptionSentences.map((sentence) =>
-                    <p key={nanoid()} className="property__text">{`${sentence}.`}</p>
-                  )}
-                </div>
-              </div>
-              <section className="property__reviews reviews">
-                <h2 className="reviews__title">
-                  Reviews &middot; <span className="reviews__amount">{roomReviews.length}</span>
-                </h2>
-                {!isAnyReviews ? `` : <ul className="reviews__list">
-                  {roomReviews.map((review) =>
-                    <li key={nanoid()} className="reviews__item">
-                      <div className="reviews__user user">
-                        <div className="reviews__avatar-wrapper user__avatar-wrapper">
-                          <img className="reviews__avatar user__avatar" src={review.avatar} width="54" height="54" alt="Reviews avatar" />
-                        </div>
-                        <span className="reviews__user-name">{review.author}</span>
-                      </div>
-                      <div className="reviews__info">
-                        <div className="reviews__rating rating">
-                          <div className="reviews__stars rating__stars">
-                            <span style={{width: `${getRatingInPercent(review.rating)}%`}}></span>
-                            <span className="visually-hidden">Rating</span>
-                          </div>
-                        </div>
-                        <p className="reviews__text">
-                          {review.text}
-                        </p>
-                        <time className="reviews__time" dateTime={formatDate(review.date)}>{review.date}</time>
-                      </div>
-                    </li>
-                  )}
-                </ul>}
-                <form className="reviews__form form" action="#" method="post">
-                  <label className="reviews__label form__label" htmlFor="review">Your review</label>
-                  <div className="reviews__rating-form form__rating">
-                    <input className="form__rating-input visually-hidden" name="rating" value="5" id="5-stars" type="radio" />
-                    <label htmlFor="5-stars" className="reviews__rating-label form__rating-label" title="perfect">
-                      <svg className="form__star-image" width="37" height="33">
-                        <use xlinkHref="#icon-star"></use>
-                      </svg>
-                    </label>
-
-                    <input className="form__rating-input visually-hidden" name="rating" value="4" id="4-stars" type="radio" />
-                    <label htmlFor="4-stars" className="reviews__rating-label form__rating-label" title="good">
-                      <svg className="form__star-image" width="37" height="33">
-                        <use xlinkHref="#icon-star"></use>
-                      </svg>
-                    </label>
-
-                    <input className="form__rating-input visually-hidden" name="rating" value="3" id="3-stars" type="radio" />
-                    <label htmlFor="3-stars" className="reviews__rating-label form__rating-label" title="not bad">
-                      <svg className="form__star-image" width="37" height="33">
-                        <use xlinkHref="#icon-star"></use>
-                      </svg>
-                    </label>
-
-                    <input className="form__rating-input visually-hidden" name="rating" value="2" id="2-stars" type="radio" />
-                    <label htmlFor="2-stars" className="reviews__rating-label form__rating-label" title="badly">
-                      <svg className="form__star-image" width="37" height="33">
-                        <use xlinkHref="#icon-star"></use>
-                      </svg>
-                    </label>
-
-                    <input className="form__rating-input visually-hidden" name="rating" value="1" id="1-star" type="radio" />
-                    <label htmlFor="1-star" className="reviews__rating-label form__rating-label" title="terribly">
-                      <svg className="form__star-image" width="37" height="33">
-                        <use xlinkHref="#icon-star"></use>
-                      </svg>
-                    </label>
-                  </div>
-                  <textarea className="reviews__textarea form__textarea" id="review" name="review" placeholder="Tell how was your stay, what you like and what can be improved"></textarea>
-                  <div className="reviews__button-wrapper">
-                    <p className="reviews__help">
-                      To submit review please make sure to set <span className="reviews__star">rating</span> and describe your stay with at least <b className="reviews__text-amount">50 characters</b>.
-                    </p>
-                    <button className="reviews__submit form__submit button" type="submit" disabled="">Submit</button>
-                  </div>
-                </form>
-              </section>
-            </div>
-          </div>
-          <section className="property__map map"></section>
-        </section>
-        <div className="container">
-          <section className="near-places places">
-            <h2 className="near-places__title">Other places in the neighbourhood</h2>
-            <div className="near-places__list places__list">
-              {neighbourhoodOffers.map((offer) =>
-                <Card
-                  key={offer.id}
-                  cardType={CardType.NEIGHBOURHOOD}
-                  offer={offer}
-                  isFavorite={favorites.includes(offer.id)}
-                />
-              )}
-            </div>
-          </section>
-        </div>
-      </main>
-    </div>
-  );
-};
-
-Room.propTypes = {
-  offerId: PropTypes.string.isRequired,
-  offers: PropTypes.arrayOf(offersPropTypes).isRequired,
-  reviews: PropTypes.arrayOf(reviewsPropTypes).isRequired,
-  favorites: PropTypes.array.isRequired,
-};
-
-export default Room;
